Replace with polyfill and converted component to directive
