+++ assets/app.css
@@ -301,3 +301,24 @@ select {
 	font-weight: bold;
 	background-color: #ffeb3b;
 }
+
+.collapsed .ngx-datatable {
+  width: 90%;
+}
+div.sidebar {
+  background: silver;
+  float: left;
+  width: 25%;
+  height: 700px;
+}
+
+
+div.sidebar.collapsed {
+  width: 5%;
+  overflow: hidden;
+}
+
+
+div.sidebar.collapsed span {
+  display: none;
+}
