+++ src/directives/index.ts
@@ -3,3 +3,4 @@ export * from './long-press.directive';
 export * from './orderable.directive';
 export * from './resizeable.directive';
 export * from './visibility.directive';
+export * from './resize-observer.directive';
